[core] fix at-everyone/at-here vulnerability, reorganise hacks
fix #254
